Fix renaming patch name in series
The condition was if line in series starts with old patch name.
However this does not catch lines starting with spaces and
catch patches that starts with that name, for instance a "foo"
patch would consider a line with "foo2" which is a mistake.
Use same code to extract patch name and "start" and "end" variables
to replace just the name of the patch in the series line.

Signed-off-by: Frediano Ziglio <frediano.ziglio@cloud.com>