Fix regression regarding GIT binary patches
Guilt already supports GIT binary patches (which are created if a file
has a .gitattribute of -diff, for instance).  But the new
extract-submodule.py helper didn't understand them, and would treat
anything after a binary patch as a patch file footer.  This would
cause it to miss submodule changes after such a patch.  When this
happened, it would print the following message to stderr:

    warning: malformed patch, treated as footer.

This commit applies the changes from commit 6cb733dc5d9ac0 of the
pdiffdiff repo to extract-submodule.py, thus adding support for GIT
binary patches.  It also adds a test case to t-072.sh to ensure that
binary patches work in Guilt.